// block-performance-1.0.0/src/library/web-client/head/type.ts
export interface Input {
    url: string;
    headers?: { name: string; value: string }[];
}

export interface OutputOk {
    status_code: number;
    status_message: string;
    headers: { name: string; value: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;